[build] move openthread-system.h under openthread
This PR moves openthread-system.h to a directory named openthread so
that projects depends on OpenThread can use it without installing
OpenThread project.
