+++ src/posix/CMakeLists.txt
@@ -31,6 +31,7 @@ set(COMMON_INCLUDES
     ${OT_PRIVATE_INCLUDES}
     ${PROJECT_SOURCE_DIR}/src/core
     ${PROJECT_SOURCE_DIR}/src/posix/platform
+    ${PROJECT_SOURCE_DIR}/src/posix/platform/include
 )
 
 if(OT_DAEMON)
