+++ src/posix/platform/platform-posix.h
@@ -43,9 +43,9 @@
 #include <sys/select.h>
 #include <sys/time.h>
 
-#include <openthread-system.h>
 #include <openthread/error.h>
 #include <openthread/instance.h>
+#include <openthread/openthread-system.h>
 
 #include "common/logging.hpp"
